refactored code to include header file. BenchMarkTestData uses constructors
